backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
bcrypt
pyjwt
Pillow
httpx
